implement the bacend api and handling the image message and the chat history
