--- .github/instructions/general-coding.instructions.md
@@ -1,27 +1,0 @@
----
-applyTo: "**"
----
-# Copilot Coding Instructions
-- Start your responses with "🤖" to indicate you're using project instructions.
-## Code Quality
-- Write clean, readable, modular code.
-- Follow existing naming and formatting styles.
-- Avoid duplicate or unused code.
-- Avoid unused or duplicated code.
-- Validate all inputs and check responses.
-- Before creating new functions or importing new packages, check the existing installed packages and local utilities in the project.
-
-
-## Error Handling
-- Use try/catch for async code.
-- Log errors with context.
-- Use React error boundaries where needed.
-
-## Data & Logic
-- use context7 MCP server when logic depends on configuration.
-- Validate all inputs and check API responses.
-
-## Best Practices
-- Prefer arrow functions and const/let.
-- Break large functions into smaller units.
-- Handle edge cases and null values gracefully.
