--- Backend/.gitignore
@@ -1,24 +1,0 @@
-.vscode/*
-!.vscode/extensions.json
-.idea
-.qodo
-.DS_Store
-*.suo
-.ntvs
-*.njsproj
-*.sln
-*.sw?
-venv
-
-# Python compiled files
-__pycache__/
-*.py[cod]
-*.pyc
-db.sqlite3
-# Media files
-*.mp4
-BehindTheSearch/static/media/
-media/
-intro.mp4
-.env
-
