--- Backend/README.md
@@ -1,7 +1,0 @@
-# ServerSideGradeProj
-
-pip install djangorestframework
-pip install google-generativeai <!--  "Optional" -->
-pip install django-cors-headers
-pip install openrouter
-pip install openai
