--- Backend/api/serializers.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from .models import ChatHistory
-
-
-class ChatHistorySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ChatHistory
-        fields = ['id', 'prompt', 'image', 'response', 'source', 'timestamp']
