+++ android/app/build.gradle.kts
@@ -8,7 +8,8 @@ plugins {
 android {
     namespace = "com.example.chatbot_app"
     compileSdk = flutter.compileSdkVersion
-    ndkVersion = flutter.ndkVersion
+    ndkVersion = "27.0.12077973" // Setting to match required plugin versions
+   // ndkVersion = flutter.ndkVersion
 
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_11
