+++ android/app/src/main/kotlin/com/front2back/chatbot_app/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.front2back.chatbot_app
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()
