+++ android/app/src/main/res/drawable-v21/launch_background.xml
