+++ android/app/src/profile/AndroidManifest.xml
