+++ pubspec.yaml
@@ -12,6 +12,7 @@ dependencies:
   http: ^1.2.0
   logger: ^2.0.2
   image_picker: ^1.0.7
+  http_parser: ^4.0.2
   path_provider: ^2.1.2
   cached_network_image: ^3.3.1
   animated_text_kit: ^4.2.2
